move job script tests to test_job_script_files
